fix(haiku): Improve type safety and error handling for clock ticks
Refactor type casts in cpu and filesystem modules for consistency. Add a robust function to convert clock ticks to nanoseconds with overflow and error checks. Update cpu_time to validate inputs and propagate errors correctly. Include tests for the new conversion functionality.

+++ include/syscape/detail/cpu/haiku.hpp
@@ -275,8 +275,8 @@ inline result<std::vector<std::uint32_t>> current_frequencies_khz() {
         if (c.current_frequency == 0) {
             return fail(errc::temporarily_unavailable);
         }
-        const auto khz =
-            static_cast<std::uint64_t>(c.current_frequency / 1000ULL);
+        const auto khz = static_cast<std::uint64_t>(
+            c.current_frequency / static_cast<std::uint64_t>(1000U));
         if (khz == 0 || khz > (std::numeric_limits<std::uint32_t>::max)()) {
             return fail(errc::malformed_data);
         }
